integrate the new deserializer
