+++ lib/transport/binary/protocol/deserializer.js
@@ -1,227 +1,537 @@
 "use strict";
 
-var RecordID = require('../../../recordid'),
-    Long = require('../../../long').Long;
-
+var RID = require('../../../recordid');
 
 /**
- * A map of start delimiters to their end delimiters.
- * @type {Object}
- */
-var START_DELIMITERS = {
-  '[': ']',
-  '{': '}',
-  '(': ')',
-  '<': '>'
-};
-
-/**
- * A map of end delimiters to their start delimiters.
- * @type {Object}
- */
-var END_DELIMITERS = {
-  ']': '[',
-  '}': '{',
-  ')': '(',
-  '>': '<'
-};
-
-/**
- * Deserialize a given serialized document.
+ * Deserialize the given record and return an object containing the values.
  *
- * @param  {String}  serialized The serialized document.
- * @param  {Object}  document   The optional document to apply the unserialized values to.
- * @param  {Boolean} isMap      If this is a map of values
- * @return {Object}             The deserialized document
+ * @param  {String} input The serialized record.
+ * @return {Object}       The deserialized record.
  */
-function deserializeDocument (serialized, document, isMap) {
-  if (serialized == null) {
-    return serialized;
-  }
-  serialized  = serialized.trim();
-  document  = document || {};
-
-  var classIndex = serialized.indexOf("@"),
-      indexOfColon  = serialized.indexOf(":"),
-      fieldIndex, field, commaIndex, value;
-
-  if (~classIndex && (indexOfColon > classIndex || !~indexOfColon)) {
-    document['@class'] = serialized.substr(0, classIndex);
-    serialized = serialized.substr(classIndex + 1);
-  }
-
-  if (!isMap) {
-    document['@type'] = 'd';
-  }
-
-  while (~(fieldIndex = serialized.indexOf(':'))) {
-    field = serialized.substr(0, fieldIndex);
-    serialized = serialized.substr(fieldIndex + 1);
-    if (field.charAt(0) === "\"" && field.charAt(field.length - 1) === "\"") {
-      field = field.substring(1, field.length - 1);
-    }
-    commaIndex = findCommaIndex(serialized);
-    value = serialized.substr(0, commaIndex);
-    serialized = serialized.substr(commaIndex + 1);
-    value = deserializeValue(value);
-    document[field] = value;
-  }
-
-  return document;
-}
-
-/**
- * Deserialize a bag of values.
- *
- * @param  {String}     value  The serialized bag.
- * @return {RecordID[]}        The unserialized bag.
- */
-function deserializeBag (value) {
-  var buffer = new Buffer(value.slice(1,-1), 'base64'),
-      offset = 0,
-      status = buffer.readUInt8(offset++),
-      total = buffer.readUInt32BE(offset),
-      content = [],
-      i, cluster, position;
-
-  offset += 4;
-  for (i = 0; i < total; i++) {
-    cluster = buffer.readInt16BE(offset);
-    offset += 2;
-    position = Long.fromBits(buffer.readUInt32BE(offset + 4), buffer.readInt32BE(offset)).toNumber();
-    offset += 8;
-    content.push(new RecordID({
-      cluster: cluster,
-      position: position
-    }));
-  }
-  return content;
-}
-
-/**
- * Deserialize a given value into the right type.
- *
- * @param  {String} value The serialized value to unserialize.
- * @return {Mixed}        The deserialized value.
- */
-function deserializeValue (value) {
-  value = (''+value).trim();
-  if(value === '') {
+function deserialize (input) {
+  var record = {'@type': 'd'},
+      chunk, key, value;
+  if (!input) {
     return null;
   }
-
-  if (value === 'true' || value === 'false') {
-    return value === 'true';
-  }
-
-  var firstChar = value.charAt(0),
-      lastChar = value.charAt(value.length - 1),
-      values, rid;
-
-  if (firstChar === '"') {
-    // string
-    return value.substring(1, value.length - 1).replace(/\\"/g, '"').replace(/\\\\/, '\\');
-  }
-  else if (firstChar === '%' && lastChar === ';') {
-    // RID bag
-    return deserializeBag(value);
-  }
-  else if (lastChar === 't' || lastChar === 'a') {
-    // date
-    return new Date(parseInt(value.substring(0, value.length)));
-  }
-  else if (firstChar === '(') {
-    // object / document
-    return deserializeDocument(value.substring(1, value.length - 1));
-  }
-  if (firstChar === '{') {
-    // map
-    return deserializeDocument(value.substring(1, value.length - 1), {}, true);
-  }
-  else if (firstChar === '[' || firstChar === '<') {
-    //process Set <...> like List [...]
-    values = splitList(value.substring(1, value.length - 1));
-    for (var i = 0, length = values.length; i < length; i++) {
-      values[i] = deserializeValue(values[i]);
-    }
-    return values;
-  }
-  else if (lastChar === 'b') {
-    // byte
-    return String.fromCharCode(+(value.substring(0, value.length)));
-  }
-  else if (lastChar === 'l' || lastChar === 's' || lastChar === 'c') {
-    // integer
-    return parseInt(value.substring(0, value.length), 10);
-  }
-  else if (lastChar === 'f' || lastChar === 'd') {
-    // float / decimal
-    return parseFloat(value.substring(0, value.length));
-  }
-  else if (+value == value) {
-    // integer
-    return +value;
-  }
-  else if (value && (rid = RecordID.parse(value))) {
-    return rid;
+  chunk = eatFirstKey(input);
+  if (chunk[2]) {
+    // this is actually a class name
+    record['@class'] = chunk[0];
+    input = chunk[1];
+    chunk = eatKey(input);
+    key = chunk[0];
+    input = chunk[1];
   }
   else {
-    return value;
+    key = chunk[0];
+    input = chunk[1];
   }
+
+  // read the first value.
+  chunk = eatValue(input);
+  value = chunk[0];
+  input = chunk[1];
+  record[key] = value;
+
+  while (input.length) {
+    if (input.charAt(0) === ',') {
+      input = input.slice(1);
+    }
+    else {
+      break;
+    }
+    chunk = eatKey(input);
+    key = chunk[0];
+    input = chunk[1];
+    if (input.length) {
+      chunk = eatValue(input);
+      value = chunk[0];
+      input = chunk[1];
+      record[key] = value;
+    }
+    else {
+      record[key] = null;
+    }
+  }
+
+  return record;
 }
 
-
 /**
- * Find the index of the first comma in the given serialized input,
- * disregarding values that appear within delimiters.
+ * Consume the first field key, which could be a class name.
  *
- * @param  {String} serialized The serialized value to inspect.
- * @return {Integer}           The index of the first comma.
+ * @param  {String}           input The input to parse.
+ * @return {[String, String]}       The collected key, and any remaining input.
  */
-function findCommaIndex (serialized) {
-  var delimiters = [],
-      current, prev, i, length;
-  for (i = 0, length = serialized.length; i < length; i++) {
-    current = serialized[i];
-    prev = delimiters[delimiters.length - 1];
-    if (current === "," && delimiters.length === 0) {
-      return i;
+function eatFirstKey (input) {
+  var length = input.length,
+      collected = '',
+      isClassName = false,
+      result, c, i;
+
+  if (input.charAt(0) === '"') {
+    result = eatString(input.slice(1));
+    return [result[0], result[1].slice(1)];
+  }
+
+  for (i = 0; i < length; i++) {
+    c = input.charAt(i);
+    if (c === '@') {
+      isClassName = true;
+      break;
     }
-    else if (START_DELIMITERS[current] && prev !== "\"") {
-      delimiters.push(current);
+    else if (c === ':') {
+      break;
     }
-    else if (END_DELIMITERS[current] && prev !== "\"" && (current === START_DELIMITERS[prev] || current === '"')) {
-      delimiters.pop();
-    }
-    else if (current === "\"" && prev === "\"" && i > 0 && serialized[i - 1] !== "\\") {
-      delimiters.pop();
-    }
-    else if (current === "\"" && prev !== "\"") {
-      delimiters.push(current);
+    else {
+      collected += c;
     }
   }
-  return serialized.length;
+
+  return [collected, input.slice(i + 1), isClassName];
+}
+
+
+/**
+ * Consume a field key, which may or may not be quoted.
+ *
+ * @param  {String}           input The input to parse.
+ * @return {[String, String]}       The collected key, and any remaining input.
+ */
+function eatKey (input) {
+  var length = input.length,
+      collected = '',
+      result, c, i;
+
+  if (input.charAt(0) === '"') {
+    result = eatString(input.slice(1));
+    return [result[0], result[1].slice(1)];
+  }
+
+  for (i = 0; i < length; i++) {
+    c = input.charAt(i);
+    if (c === ':') {
+      break;
+    }
+    else {
+      collected += c;
+    }
+  }
+
+  return [collected, input.slice(i + 1)];
+}
+
+
+
+/**
+ * Consume a field value.
+ *
+ * @param  {String}           input The input to parse.
+ * @return {[Mixed, String]}        The collected value, and any remaining input.
+ */
+function eatValue (input) {
+  var c, n;
+  c = input.charAt(0);
+  while (c === ' ' && input.length) {
+    input = input.slice(1);
+    c = input.charAt(0);
+  }
+
+  if (!input.length || c === ',') {
+    // this is a null field.
+    return [null, input];
+  }
+  else if (c === '"') {
+    return eatString(input.slice(1));
+  }
+  else if (c === '#') {
+    return eatRID(input.slice(1));
+  }
+  else if (c === '[') {
+    return eatArray(input.slice(1));
+  }
+  else if (c === '<') {
+    return eatSet(input.slice(1));
+  }
+  else if (c === '{') {
+    return eatMap(input.slice(1));
+  }
+  else if (c === '(') {
+    return eatRecord(input.slice(1));
+  }
+  else if (c === '%') {
+    return eatBag(input.slice(1));
+  }
+  else if (c === '_') {
+    return eatBinary(input.slice(1));
+  }
+  else if (c === '-' || c === '0' || +c) {
+    return eatNumber(input);
+  }
+  else if (c === 'n' && input.slice(0, 4) === 'null') {
+    return [null, input.slice(4)];
+  }
+  else if (c === 't' && input.slice(0, 4) === 'true') {
+    return [true, input.slice(4)];
+  }
+  else if (c === 'f' && input.slice(0, 5) === 'false') {
+    return [false, input.slice(5)];
+  }
+  else {
+    return [null, input];
+  }
 }
 
 /**
- * Split a serialized list into items that can then be parsed individually.
- * @param  {String} value  The list value to split.
- * @return {String[]}      The split items, ready to be parsed
+ * Consume a string
+ *
+ * @param  {String}           input The input to parse.
+ * @return {[String, String]}       The collected string, and any remaining input.
  */
-function splitList (value) {
-  var result = [],
-      commaAt;
-  while (value.length) {
-    commaAt = findCommaIndex(value);
-    result.push(value.substr(0, commaAt));
-    value = value.substr(commaAt + 1);
+function eatString (input) {
+  var length = input.length,
+      collected = '',
+      c, i;
+
+  for (i = 0; i < length; i++) {
+    c = input.charAt(i);
+    if (c === '\\') {
+      // escape, skip to the next character
+      i++;
+      collected += input.charAt(i);
+      continue;
+    }
+    else if (c === '"') {
+      break;
+    }
+    else {
+      collected += c;
+    }
   }
-  return result;
+
+  return [collected, input.slice(i + 1)];
+}
+
+/**
+ * Consume a number.
+ *
+ * If the number has a suffix, consume it also and instantiate the right type, e.g. for dates
+ *
+ * @param  {String}           input The input to parse.
+ * @return {[Mixed, String]}        The collected number, and any remaining input.
+ */
+function eatNumber (input) {
+  var length = input.length,
+      collected = '',
+      c, i;
+
+  for (i = 0; i < length; i++) {
+    c = input.charAt(i);
+    if (c === '-' || c === '.' || c === '0' || +c) {
+      collected += c;
+    }
+    else {
+      break;
+    }
+  }
+
+  collected = +collected;
+  input = input.slice(i);
+
+  c = input.charAt(0);
+
+  if (c === 'a' || c === 't') {
+    collected = new Date(collected);
+    input = input.slice(1);
+  }
+  else if (c === 'b' || c === 's' || c === 'l' || c === 'f' || c == 'd' || c === 'c') {
+    input = input.slice(1);
+  }
+
+  return [collected, input];
+}
+
+/**
+ * Consume a Record ID.
+ *
+ * @param  {String}           input The input to parse.
+ * @return {[RID, String]}          The collected record id, and any remaining input.
+ */
+function eatRID (input) {
+  var length = input.length,
+      collected = '',
+      cluster, c, i;
+
+  for (i = 0; i < length; i++) {
+    c = input.charAt(i);
+    if (cluster === undefined && c === ':') {
+      cluster = +collected;
+      collected = '';
+    }
+    else if (c === '0' || +c) {
+      collected += c;
+    }
+    else {
+      break;
+    }
+  }
+
+  return [new RID({cluster: cluster, position: +collected}), input.slice(i)];
 }
 
 
-// export the public methods
+/**
+ * Consume an array.
+ *
+ * @param  {String}           input   The input to parse.
+ * @return {[Array, String]}          The collected array, and any remaining input.
+ */
+function eatArray (input) {
+  var length = input.length,
+      array = [],
+      chunk, c;
+
+  while (input.length) {
+    c = input.charAt(0);
+    if (c === ',') {
+      input = input.slice(1);
+    }
+    else if (c === ']') {
+      input = input.slice(1);
+      break;
+    }
+    chunk = eatValue(input);
+    array.push(chunk[0]);
+    input = chunk[1];
+  }
+  return [array, input];
+}
 
 
-exports.deserializeValue = deserializeValue;
-exports.deserializeDocument = deserializeDocument;
+/**
+ * Consume a set.
+ *
+ * @param  {String}           input   The input to parse.
+ * @return {[Array, String]}          The collected set, and any remaining input.
+ */
+function eatSet (input) {
+  var length = input.length,
+      set = [],
+      chunk, c;
+
+  while (input.length) {
+    c = input.charAt(0);
+    if (c === ',') {
+      input = input.slice(1);
+    }
+    else if (c === '>') {
+      input = input.slice(1);
+      break;
+    }
+    chunk = eatValue(input);
+    set.push(chunk[0]);
+    input = chunk[1];
+  }
+
+  return [set, input];
+}
+
+/**
+ * Consume a map (object).
+ *
+ * @param  {String}           input   The input to parse.
+ * @return {[Object, String]}         The collected map, and any remaining input.
+ */
+function eatMap (input) {
+  var length = input.length,
+      map = {},
+      key, value, chunk, c;
+
+  while (input.length) {
+    c = input.charAt(0);
+    if (c === ' ') {
+      input = input.slice(1);
+      continue;
+    }
+    if (c === ',') {
+      input = input.slice(1);
+    }
+    else if (c === '}') {
+      input = input.slice(1);
+      break;
+    }
+
+    chunk = eatKey(input);
+    key = chunk[0];
+    input = chunk[1];
+    if (input.length) {
+      chunk = eatValue(input);
+      value = chunk[0];
+      input = chunk[1];
+      map[key] = value;
+    }
+    else {
+      map[key] = null;
+    }
+  }
+
+  return [map, input];
+}
+
+/**
+ * Consume an embedded record.
+ *
+ * @param  {String}           input   The input to parse.
+ * @return {[Object, String]}         The collected record, and any remaining input.
+ */
+function eatRecord (input) {
+  var record = {'@type': 'd'},
+      chunk, c, key, value;
+
+  while (input.length) {
+    c = input.charAt(0);
+    if (c === ' ') {
+      input = input.slice(1);
+      continue;
+    }
+    else if (c === ')') {
+      // empty record.
+      input = input.slice(1);
+      return [record, input];
+    }
+    else {
+      break;
+    }
+  }
+
+  chunk = eatFirstKey(input);
+
+  if (chunk[2]) {
+    // this is actually a class name
+    record['@class'] = chunk[0];
+    input = chunk[1];
+    chunk = eatKey(input);
+    while (input.length) {
+      c = input.charAt(0);
+      if (c === ' ') {
+        input = input.slice(1);
+        continue;
+      }
+      else if (c === ')') {
+        // empty record.
+        input = input.slice(1);
+        return [record, input];
+      }
+      else {
+        break;
+      }
+    }
+    key = chunk[0];
+    input = chunk[1];
+  }
+  else {
+    key = chunk[0];
+    input = chunk[1];
+  }
+
+  // read the first value.
+  chunk = eatValue(input);
+  value = chunk[0];
+  input = chunk[1];
+  record[key] = value;
+
+  while (input.length) {
+    c = input.charAt(0);
+    if (c === ' ') {
+      input = input.slice(1);
+      continue;
+    }
+    if (c === ',') {
+      input = input.slice(1);
+    }
+    else if (c === ')') {
+      input = input.slice(1);
+      break;
+    }
+    chunk = eatKey(input);
+    key = chunk[0];
+    input = chunk[1];
+    if (input.length) {
+      chunk = eatValue(input);
+      value = chunk[0];
+      input = chunk[1];
+      record[key] = value;
+    }
+    else {
+      record[key] = null;
+    }
+  }
+
+  return [record, input];
+}
+
+/**
+ * Consume a RID Bag.
+ *
+ * @param  {String}            input   The input to parse.
+ * @return {[Object, String]}          The collected bag, and any remaining input.
+ */
+function eatBag (input) {
+  var length = input.length,
+      collected = '',
+      i, bag, chunk, c;
+
+  for (i = 0; i < length; i++) {
+    c = input.charAt(i);
+    if (c === ';') {
+      break;
+    }
+    else {
+      collected += c;
+    }
+  }
+  input = input.slice(i + 1);
+
+  return [collected, input];
+}
+
+
+/**
+ * Consume a binary buffer.
+ *
+ * @param  {String}            input   The input to parse.
+ * @return {[Object, String]}          The collected bag, and any remaining input.
+ */
+function eatBinary (input) {
+  var length = input.length,
+      collected = '',
+      i, bag, chunk, c;
+
+  for (i = 0; i < length; i++) {
+    c = input.charAt(i);
+    if (c === '_' || c === ',' || c === ')' || c === '>' || c === '}' || c === ']') {
+      break;
+    }
+    else {
+      collected += c;
+    }
+  }
+  input = input.slice(i + 1);
+
+  return [collected, input];
+}
+
+
+
+exports.deserialize = deserialize;
+exports.eatKey = eatKey;
+exports.eatValue = eatValue;
+exports.eatString = eatString;
+exports.eatNumber = eatNumber;
+exports.eatRID = eatRID;
+exports.eatArray = eatArray;
+exports.eatSet = eatSet;
+exports.eatMap = eatMap;
+exports.eatRecord = eatRecord;
+exports.eatBag = eatBag;
