+++ lib/transport/binary/protocol/operation.js
@@ -276,7 +276,7 @@ Operation.prototype.readError = function (fieldName, reader) {
  */
 Operation.prototype.readObject = function (fieldName, reader) {
   this.readOps.push(['String', [fieldName, function (data, fieldName) {
-    data[fieldName] = deserializer.deserializeValue('{'+data[fieldName]+'}');
+    data[fieldName] = deserializer.deserialize(data[fieldName]);
     if (reader) {
       reader.call(this, data, fieldName);
     }
@@ -664,7 +664,7 @@ Operation.prototype.parseRecord = function (buffer, offset, context, fieldName, 
       .readLong('position')
       .readInt('version')
       .readString('value', function (data, key) {
-        data[key] = deserializer.deserializeDocument(data[key]);
+        data[key] = deserializer.deserialize(data[key]);
         this.stack.pop();
         this.readOps.push(function () {
           if (reader) {
@@ -838,7 +838,7 @@ Operation.prototype.parsePushedData = function (buffer, offset, context, fieldNa
   asString = buffer.toString('utf8', offset, offset + length);
   switch (asString.charAt(0)) {
     case 'R':
-      context[fieldName] = deserializer.deserializeValue('{' + asString.slice(1) + '}');
+      context[fieldName] = deserializer.deserialize(asString.slice(1));
       break;
     default:
       console.log('unsupported pushed data format: ' + asString);
