+++ lib/transport/binary/protocol/operations/command.js
@@ -3,7 +3,6 @@
 var Operation = require('../operation'),
     constants = require('../constants'),
     serializer = require('../serializer'),
-    deserializer = require('../deserializer'),
     writer = require('../writer');
 
 module.exports = Operation.extend({
