+++ lib/transport/binary/protocol/operations/record-load.js
@@ -78,7 +78,7 @@ module.exports = Operation.extend({
             data.cluster = this.data.cluster;
             data.position = this.data.position;
             if (data[fieldName] === 'd') {
-              data.content = deserializer.deserializeValue('{' + data.content + '}');
+              data.content = deserializer.deserialize(data.content);
             }
             this.stack.pop();
             this.readPayload(records, ender);
@@ -104,7 +104,7 @@ module.exports = Operation.extend({
                 .readInt('version')
                 .readString('content', function (data, fieldName) {
                   if (data.type === 'd') {
-                    data.content = deserializer.deserializeValue('{' + data.content + '}');
+                    data.content = deserializer.deserialize(data.content);
                   }
                   this.stack.pop();
                   this.readPayload(records, ender);
